requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy>=1.26.0
matplotlib==3.7.2
seaborn==0.13.0
scipy>=1.11.0
duckdb==0.9.2
httpfs
python-multipart==0.0.6
Pillow==10.1.0
lxml==4.9.3
plotly==5.17.0
scikit-learn==1.3.2
python-dotenv==1.0.0
# Enhanced dependencies for multi-modal processing
PyPDF2==3.0.1
pdfplumber==0.10.3
opencv-python==4.8.1.78
pytesseract==0.3.10
SpeechRecognition==3.10.0
pydub==0.25.1
transformers>=4.35.0
torch>=2.1.0
aiofiles==23.2.1
setuptools>=68.0.0